implemented parsing manifest.sii of workshop mods, did some refactoring
