+++ package.json
@@ -7,10 +7,14 @@
     "license": "GPL-3.0-only",
     "description": "Checks if zip archives in a given folder contain the same files - intended for checking if installed ETS2/ATS mods have conflicts",
     "scripts": {
+        "check:exe": "npm run build && smcc.exe",
         "check": "node index.js",
         "build": "pkg --config pkg-config.json index.js -o smcc"
     },
     "devDependencies": {
         "pkg": "^5.8.1"
+    },
+    "dependencies": {
+        "yauzl": "^3.2.0"
     }
 }
