+++ src/analysis.js
@@ -1,3 +1,111 @@
+const { execute7zip } = require('./seven-zip/index.js')
+
+/**
+ * @param {string} pathToArchive
+ * 
+ * @returns {Promise<string[]>} list of files in archive
+ */
+const _listFilesOfArchive = (pathToArchive, withAutomatFiles = false) => {
+    let output = ''
+
+    try {
+        output = execute7zip([
+            "l",
+            "-ba",
+            pathToArchive
+        ])
+    } catch (error) {
+        throw error
+    }
+
+    // 7zip uses the following file attributes:
+    // D.... - directory
+    // .R... - read-only
+    // ..H.. - hidden
+    // ...S. - system (windows system file)
+    // ....A - archive (normal file)
+    // an entry looks like this:
+    // date - time - file attributes - size uncompressed - size compressed - path
+    // 2025-06-02 19:16:52 ....A      1398256         3082  vehicle/truck/upgrade/paintjob/scania.t/airbrush/accessories/acc_0.dds
+    const pathList = []
+
+    output.trim().split('\n')
+    .forEach(entry => { 
+        // collapse whitespace to single spaces
+        entry = entry.trim().replaceAll(/\s+/g, ' ')
+
+        // check for control/non-printable characters
+        // and clean them up
+        if (entry.match(/[\x00-\x1F\x7F]/g)) {
+            entry = entry.replace(/[\x00-\x1F\x7F]/g, '').replace('undefined', '').replace(/0M Scan/g, '').trim()
+        }
+
+        // split the string at the spaces
+        // and check if the third entry (file attr.) do not contain a D
+        let [ _date, _time, attributes, _sizeUncomp, _sizeComp, filePath, ...rest ] = entry.split(' ')
+
+        // if there are file names with whitespaces, we'll get them in the rest param
+        // for sake of the comparison, we can just stitch them togehter with a delimiter
+        if (rest?.length > 0) {
+            filePath += rest.join('_')
+        }
+
+        // only get files not directories
+        // filePath.includes('/') || filePath.includes('\\')) -> only get files beyond the first level
+        // since the root dir only contains meta data that is not mounted into the games directories
+        if (!attributes.includes('D')
+            && (filePath.includes('/') || filePath.includes('\\'))
+            && (withAutomatFiles || !filePath.startsWith('automat'))
+        ) {
+            pathList.push(filePath)
+        }
+    })
+
+    return pathList
+}
+
+/**
+ * @param {ModArchive[]} archivesToCheck list of archive paths
+ * 
+ * @returns {Promise<MoData>}
+ */
+const gatherModDataFromArchives = (archivesToCheck, withAutomatFiles = false) => {
+    const mods      = []
+    const modErrors = []
+
+    for (const archive of archivesToCheck) {
+        const modName = archive.modName
+
+        try {
+            const pathList = _listFilesOfArchive(archive.path, withAutomatFiles)
+
+            /** @type {Mod} */
+            const mod = {
+                modName,
+                files: pathList
+            }
+
+            mods.push(mod)
+        } catch (error) {
+            /** @type {ModError} */
+            const modError = {
+                modName,
+                error
+            }
+
+            modErrors.push(modError)
+        }
+    }
+
+    /** @type {ModData} */
+    const result = {
+        mods,
+        modErrors
+    }
+
+    return result
+}
+
 /**
  * @param {ModData} modsData
  * 
@@ -43,5 +151,6 @@ const analyseModsContent = async (modsData) => {
 }
 
 module.exports = {
-    analyseModsContent
+    analyseModsContent,
+    gatherModDataFromArchives
 }
